telegram-bot/requirements.txt
python-telegram-bot==20.7
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
pytest==8.4.0
pytest-asyncio==1.0.0
pydantic==2.5.0
python-multipart==0.0.19
python-dotenv==1.0.0
requests==2.32.4
